chore(config): delete settled experimental flags and dead GHG/waste fields
