--- bedrock/utils/config/configs/2025_usa_cornerstone_full_model_scaling_for_A_and_B.yaml
@@ -1,18 +1,0 @@
-# Fully implemented Cornerstone 2025 model with addition of 
-# Use Cornerstone 2023 GHG FBS (GHG_national_Cornerstone_2023) via new_ghg_method
-# Pipeline loads E from flowsa.
-# Transform B matrix with USEEIO method
-# Implement waste disaggregation.
-
-#####
-# Methodology selection
-#####
-use_cornerstone_2026_model_schema: True
-new_ghg_method: true
-use_E_data_year_for_x_in_B: True
-implement_waste_disaggregation: True
-cornerstone_industry_avg_margins: True
-scale_a_matrix_with_summary_tables: True
-adjust_summary_A_and_q_dollar_year: True
-use_scaled_x_and_scaled_Vnorm_for_B: True #unique setting not in cornerstone full model A summary realigned
-# or the full model
