--- bedrock/utils/config/configs/reconciling_data_years/model1.yaml
@@ -1,9 +1,0 @@
-## Config for model 1 in Reconciling Data Years study
-# Use data from last common reporting year (2017)
-
-use_cornerstone_2026_model_schema: true
-implement_waste_disaggregation: true 
-model_base_year: 2017
-usa_ghg_data_year: 2017
-scale_a_matrix_with_useeio_method: true #necessary not to inflate or scale A
-new_ghg_method: true #get FBS from cache
