--- bedrock/utils/config/configs/reconciling_data_years/model3a.yaml
@@ -1,13 +1,0 @@
-## Config for model 3 in Reconciling Data Years study
-# Use IOT as is to make the A matrix in 2017 dollars and use annual GHG data with annual output to make the B matrix; then deflate the B matrix to IOT year (2017)
-
-model_base_year: 2017
-use_cornerstone_2026_model_schema: true
-implement_waste_disaggregation: true
-deflate_x_to_detail_io_year_for_B: true
-use_E_data_year_for_x_in_B: true
-scale_a_matrix_with_useeio_method: true #necessary not to inflate or scale A
-get_q_from_authoritative_x: true
-apply_inflation_to_V: false
-new_ghg_method: true #get FBS from cache
-
