+++ bedrock/utils/config/configs/test_usa_config_waste_disagg_electricity.yaml
@@ -3,8 +3,3 @@ usa_ghg_data_year: 2023
 use_cornerstone_2026_model_schema: True
 implement_waste_disaggregation: True
 implement_electricity_reallocation: True
-eeio_waste_disaggregation:
-  use_weights_file: "extract/disaggregation/waste_disagg_inputs/WasteDisaggregationDetail2017_Use.csv"
-  make_weights_file: "extract/disaggregation/waste_disagg_inputs/WasteDisaggregationDetail2017_Make.csv"
-  year: 2017
-  source_name: "WasteDisaggregationDetail2017"
