+++ bedrock/utils/config/configs/useeio_phoebe_23.yaml
@@ -27,7 +27,6 @@ scale_a_matrix_with_useeio_method: true
 use_useeio_schema: true
 deflate_x_to_detail_io_year_for_B: true
 use_ghg_national_2023_m2: true
-skip_scrap_adjustment_in_vnorm: true
 
 # Must remain false whenever use_ghg_national_2023_m2 is true.
 new_ghg_method: false
