+++ bedrock/utils/config/configs/v03_waterfall_useeio_g1_schema_ghg.yaml
@@ -15,7 +15,6 @@ scale_a_matrix_with_useeio_method: true
 use_useeio_schema: false
 deflate_x_to_detail_io_year_for_B: true
 use_ghg_national_2023_m2: false
-skip_scrap_adjustment_in_vnorm: true
 new_ghg_method: true
 
 useeio_margins: true
