+++ bedrock/utils/validation/__tests__/test_diagnostics_helpers.py
@@ -58,7 +58,7 @@ class TestDNNewInflatedEligibility:
         )
         ok, reason = d_n_new_inflated_eligibility(cfg)
         assert ok is False
-        assert 'double-apply' in reason
+        assert 'no denominator inflation adjustment' in reason
 
     def test_skip_legacy_non_cornerstone(self) -> None:
         cfg = USAConfig(
